refactor code in feature_extraction and similarity_search modules
