+++ fis/feature_extraction/detection/base.py
